+++ src/vector.h
@@ -7,14 +7,14 @@
 #define VECTOR_INIT  20
 
 typedef struct  {
-    uint32_t *vect;
+    uint64_t*vect;
     int size;
     int tos;
 } vector;
 
 vector* vector_init();
-void vector_push_back(vector *v, uint32_t value);
+void vector_push_back(vector *v, uint64_t value);
 void vector_free(vector *v);
-int vector_find(vector *v, uint32_t value);
+int vector_find(vector *v, uint64_t value);
 
 #endif //MCA_VECTOR_H
